requirements.txt
Django==1.10.6
pymongo==2.8
mongoengine==0.9.0
djangorestframework==3.6.2
django-rest-framework-mongoengine==3.3.1
